Move RoleTag to the types folder
