+++ ui/src/ModalFormComponents/AddNewTagRow.tsx
@@ -20,7 +20,7 @@ import React, {useState} from 'react';
 import {createDataTestId} from '../Utils/ReactUtils';
 import EditTagRow from './EditTagRow';
 import {TagRequest} from '../Tags/TagRequest.interface';
-import {Color} from '../Roles/RoleTag.interface';
+import {Color} from 'Types/RoleTag';
 import {TagNameType, TagType} from '../SortingAndFiltering/FilterLibraries';
 
 interface Props {
